requirements.txt
streamlit>=1.36
pandas>=2.1
gspread>=6.0.0
google-auth>=2.29.0
